pyproject.toml
[project]
name = "karpov-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bert-score>=0.3.13",
    "bs4>=0.0.2",
    "faiss-cpu>=1.9.0.post1",
    "ipykernel>=6.29.5",
    "lxml>=5.3.0",
    "openai>=1.60.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.41.1",
]
